- Replaced UIWebView with WKWebview. - Updated demo project to support swift 5.1 (Xcode 11.2.1). - Organized demo projects storyboards into one file. - Cleaned up linting / formating issues. - Fixed navigation issues with demo project.
