+++ Demo/Controllers/ViewController.swift
@@ -10,28 +10,18 @@ import UIKit
 import PDFGenerator
 
 class ViewController: UIViewController {
-    
+
     fileprivate var outputAsData: Bool = false
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
 
     fileprivate func getImagePath(_ number: Int) -> String {
         return Bundle.main.path(forResource: "sample_\(number)", ofType: "jpg")!
     }
-    
+
     fileprivate func getDestinationPath(_ number: Int) -> String {
         return NSHomeDirectory() + "/sample\(number).pdf"
     }
-    
-    @objc @IBAction fileprivate func generateSamplePDFFromViews(_ sender: AnyObject?) {
+
+    @IBAction fileprivate func generateSamplePDFFromViews(_ sender: AnyObject?) {
         let v1 = UIScrollView(frame: CGRect(x: 0, y: 0, width: 100, height: 100))
         let v2 = UIView(frame: CGRect(x: 0, y: 0, width: 100, height: 200))
         let v3 = UIView(frame: CGRect(x: 0, y: 0, width: 100, height: 200))
@@ -39,7 +29,7 @@ class ViewController: UIViewController {
         v1.contentSize = CGSize(width: 100, height: 100)
         v2.backgroundColor = UIColor.green
         v3.backgroundColor = UIColor.blue
-        
+
         do {
             let dst = getDestinationPath(1)
             if outputAsData {
@@ -53,8 +43,8 @@ class ViewController: UIViewController {
             print(e)
         }
     }
-    
-    @objc @IBAction fileprivate func generateSamplePDFFromImages(_ sender: AnyObject?) {
+
+    @IBAction fileprivate func generateSamplePDFFromImages(_ sender: AnyObject?) {
         let dst = getDestinationPath(2)
         autoreleasepool {
             do {
@@ -74,8 +64,8 @@ class ViewController: UIViewController {
             }
         }
     }
-    
-    @objc @IBAction fileprivate func generateSamplePDFFromImagePaths(_ sender: AnyObject?) {
+
+    @IBAction fileprivate func generateSamplePDFFromImagePaths(_ sender: AnyObject?) {
         do {
             let dst = getDestinationPath(3)
             var imagePaths = [String]()
@@ -93,13 +83,13 @@ class ViewController: UIViewController {
             print(e)
         }
     }
-    
-    @objc @IBAction fileprivate func generateSamplePDFFromPages(_ sender: AnyObject?) {
+
+    @IBAction fileprivate func generateSamplePDFFromPages(_ sender: AnyObject?) {
         let v1 = UIView(frame: CGRect(x: 0, y: 0, width: 100, height: 100))
         v1.backgroundColor = UIColor.red
         let v2 = UIView(frame: CGRect(x: 0, y: 0, width: 100, height: 200))
         v2.backgroundColor = UIColor.green
-        
+
         let page1 = PDFPage.view(v1)
         let page2 = PDFPage.view(v2)
         let page3 = PDFPage.whitePage(CGSize(width: 200, height: 100))
@@ -117,35 +107,17 @@ class ViewController: UIViewController {
             openPDFViewer(dst)
         } catch let e {
             print(e)
-
         }
-    }
-    
-    @objc @IBAction fileprivate func generatePDFFromStackedScrollView(_: AnyObject?) {
-        let storyboard = UIStoryboard(name: "PDFOutput", bundle: nil)
-        let vc = storyboard.instantiateInitialViewController()!
-        
-        present(vc, animated: true, completion: nil)
     }
 
     fileprivate func openPDFViewer(_ pdfPath: String) {
-        let url = URL(fileURLWithPath: pdfPath)        
-        let storyboard = UIStoryboard(name: "PDFPreviewVC", bundle: nil)
-        let vc = storyboard.instantiateInitialViewController() as! PDFPreviewVC
-        vc.setupWithURL(url)
-        present(vc, animated: true, completion: nil)
+        self.performSegue(withIdentifier: "PreviewVC", sender: pdfPath)
     }
 
-    @objc @IBAction fileprivate func goSampleTableView(_ sender: AnyObject?) {
-        let storyboard = UIStoryboard(name: "SampleTableViewController", bundle: nil)
-        let vc = storyboard.instantiateInitialViewController() as! SampleTableViewController
-        present(vc, animated: true, completion: nil)
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        if let pdfPreviewVC = segue.destination as? PDFPreviewVC, let pdfPath = sender as? String {
+            let url = URL(fileURLWithPath: pdfPath)
+            pdfPreviewVC.setupWithURL(url)
+        }
     }
-    
-    @objc @IBAction fileprivate func goSampleWebView(_ sender: AnyObject?) {
-        let storyboard = UIStoryboard(name: "WebViewController", bundle: nil)
-        let vc = storyboard.instantiateInitialViewController() as! WebViewController
-        present(vc, animated: true, completion: nil)
-    }
-
 }
