--- Demo/PDFPreviewVC.storyboard
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="9531" systemVersion="14F27" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" initialViewController="7Hn-bD-ujW">
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="9529"/>
-    </dependencies>
-    <scenes>
-        <!--PreviewVC-->
-        <scene sceneID="BoV-cP-0sY">
-            <objects>
-                <viewController id="7Hn-bD-ujW" customClass="PDFPreviewVC" customModule="Demo" customModuleProvider="target" sceneMemberID="viewController">
-                    <layoutGuides>
-                        <viewControllerLayoutGuide type="top" id="5yI-p4-Eij"/>
-                        <viewControllerLayoutGuide type="bottom" id="hAG-XP-Aaz"/>
-                    </layoutGuides>
-                    <view key="view" contentMode="scaleToFill" id="YGd-n0-0Xl">
-                        <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="PdX-6V-CI9">
-                                <rect key="frame" x="0.0" y="0.0" width="600" height="64"/>
-                                <subviews>
-                                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="Mrg-Ib-dGc">
-                                        <rect key="frame" x="16" y="27" width="60" height="30"/>
-                                        <constraints>
-                                            <constraint firstAttribute="height" constant="30" id="XeW-vz-XoI"/>
-                                            <constraint firstAttribute="width" constant="60" id="mBz-Nb-mFc"/>
-                                        </constraints>
-                                        <state key="normal" title="Close"/>
-                                        <connections>
-                                            <action selector="close:" destination="7Hn-bD-ujW" eventType="touchUpInside" id="nNU-hg-ofw"/>
-                                        </connections>
-                                    </button>
-                                </subviews>
-                                <color key="backgroundColor" red="0.93122661110000005" green="0.93119871620000005" blue="0.9312145114" alpha="1" colorSpace="calibratedRGB"/>
-                                <constraints>
-                                    <constraint firstItem="Mrg-Ib-dGc" firstAttribute="centerY" secondItem="PdX-6V-CI9" secondAttribute="centerY" constant="10" id="7La-5g-lE7"/>
-                                    <constraint firstItem="Mrg-Ib-dGc" firstAttribute="leading" secondItem="PdX-6V-CI9" secondAttribute="leading" constant="16" id="Ypn-6q-kgk"/>
-                                    <constraint firstAttribute="height" constant="64" id="fLW-s8-iTd"/>
-                                </constraints>
-                            </view>
-                            <webView contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Igd-XZ-bT9">
-                                <rect key="frame" x="0.0" y="64" width="600" height="536"/>
-                                <color key="backgroundColor" red="0.36078431370000003" green="0.38823529410000002" blue="0.4039215686" alpha="1" colorSpace="deviceRGB"/>
-                            </webView>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                        <constraints>
-                            <constraint firstItem="PdX-6V-CI9" firstAttribute="top" secondItem="YGd-n0-0Xl" secondAttribute="top" id="5fT-BR-KEd"/>
-                            <constraint firstItem="Igd-XZ-bT9" firstAttribute="leading" secondItem="YGd-n0-0Xl" secondAttribute="leading" id="AHt-ne-34F"/>
-                            <constraint firstAttribute="trailing" secondItem="Igd-XZ-bT9" secondAttribute="trailing" id="Fbl-js-706"/>
-                            <constraint firstItem="Igd-XZ-bT9" firstAttribute="top" secondItem="PdX-6V-CI9" secondAttribute="bottom" id="L1K-sb-l9K"/>
-                            <constraint firstAttribute="trailing" secondItem="PdX-6V-CI9" secondAttribute="trailing" id="NDT-Na-kEi"/>
-                            <constraint firstItem="PdX-6V-CI9" firstAttribute="leading" secondItem="YGd-n0-0Xl" secondAttribute="leading" id="RhG-Ek-Gxm"/>
-                            <constraint firstAttribute="bottom" secondItem="Igd-XZ-bT9" secondAttribute="bottom" id="dVl-ig-xU2"/>
-                        </constraints>
-                    </view>
-                    <connections>
-                        <outlet property="webView" destination="Igd-XZ-bT9" id="ID4-Bg-EVs"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="1cM-d1-vR2" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="684" y="212"/>
-        </scene>
-    </scenes>
-</document>
