--- Demo/SampleTableViewController.storyboard
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="9532" systemVersion="15D21" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" initialViewController="4wT-3U-j3F">
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="9530"/>
-        <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
-    </dependencies>
-    <scenes>
-        <!--Sample Table View Controller-->
-        <scene sceneID="TKp-H0-I3P">
-            <objects>
-                <tableViewController id="4wT-3U-j3F" customClass="SampleTableViewController" customModule="Demo" customModuleProvider="target" sceneMemberID="viewController">
-                    <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="44" sectionHeaderHeight="28" sectionFooterHeight="28" id="Gfi-OA-RfQ">
-                        <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                        <prototypes>
-                            <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="Cell" id="hRd-LY-XjG" customClass="SampleTableViewCell" customModule="Demo" customModuleProvider="target">
-                                <rect key="frame" x="0.0" y="28" width="600" height="44"/>
-                                <autoresizingMask key="autoresizingMask"/>
-                                <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="hRd-LY-XjG" id="cFY-ba-p1v">
-                                    <rect key="frame" x="0.0" y="0.0" width="600" height="43"/>
-                                    <autoresizingMask key="autoresizingMask"/>
-                                    <subviews>
-                                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="BHf-3f-fHF">
-                                            <rect key="frame" x="16" y="11" width="200" height="21"/>
-                                            <constraints>
-                                                <constraint firstAttribute="height" constant="21" id="3lx-Yu-VcB"/>
-                                                <constraint firstAttribute="width" constant="200" id="yTE-Gm-4vg"/>
-                                            </constraints>
-                                            <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                            <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
-                                            <nil key="highlightedColor"/>
-                                        </label>
-                                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="right" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="SYO-87-FkQ">
-                                            <rect key="frame" x="384" y="11" width="200" height="21"/>
-                                            <constraints>
-                                                <constraint firstAttribute="width" constant="200" id="EC5-Ob-cwi"/>
-                                                <constraint firstAttribute="height" constant="21" id="Ki8-nY-Zw3"/>
-                                            </constraints>
-                                            <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                            <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
-                                            <nil key="highlightedColor"/>
-                                        </label>
-                                    </subviews>
-                                    <constraints>
-                                        <constraint firstItem="BHf-3f-fHF" firstAttribute="leading" secondItem="cFY-ba-p1v" secondAttribute="leadingMargin" constant="8" id="OWc-d2-3Kc"/>
-                                        <constraint firstItem="BHf-3f-fHF" firstAttribute="centerY" secondItem="cFY-ba-p1v" secondAttribute="centerY" id="apx-WT-hCh"/>
-                                        <constraint firstItem="SYO-87-FkQ" firstAttribute="centerY" secondItem="cFY-ba-p1v" secondAttribute="centerY" id="ix7-6X-G5F"/>
-                                        <constraint firstAttribute="trailingMargin" secondItem="SYO-87-FkQ" secondAttribute="trailing" constant="8" id="mKQ-oZ-wJR"/>
-                                    </constraints>
-                                </tableViewCellContentView>
-                                <connections>
-                                    <outlet property="leftLabel" destination="BHf-3f-fHF" id="DrU-nI-LuS"/>
-                                    <outlet property="rightLabel" destination="SYO-87-FkQ" id="cDo-TB-Tml"/>
-                                </connections>
-                            </tableViewCell>
-                        </prototypes>
-                        <connections>
-                            <outlet property="dataSource" destination="4wT-3U-j3F" id="4rD-nT-RaB"/>
-                            <outlet property="delegate" destination="4wT-3U-j3F" id="Zpt-pd-xjw"/>
-                        </connections>
-                    </tableView>
-                </tableViewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="gDh-v1-sOp" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="625" y="418"/>
-        </scene>
-    </scenes>
-</document>
